Add dynamic can id support
This enables the boards (probe/fan boards) to request a can id from the
master. This can id gets used to generate (e.g. by adding +1 for responses)
a range in which the boards operate. The format of the commands has also
changed, since they now support dynamic ranges too.
Comments where added on a huge scale and in general this commit is way
to huge… Anyway, it fixes a lot of stuff!